backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0
geoalchemy2==0.17.1
pydantic[email]==2.10.3
pydantic-settings==2.7.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
celery==5.4.0
redis==5.2.1
httpx==0.28.1
Pillow==11.1.0
exifread==3.0.0
aiofiles==24.1.0
slowapi==0.1.9
python-dotenv==1.0.1
boto3==1.36.1
# AI Model — YOLOv8 inference
ultralytics==8.3.106
opencv-python-headless==4.11.0.86
torch==2.6.0
torchvision==0.21.0
